// src/app/modules/deliverers/enums/deliverer.enum.ts
export enum DelivererActionTypes {
  FETCH_DELIVERER = '[DELIVERERS] FETCH_DELIVERER',
  FETCH_DELIVERER_SUCCESS = '[DELIVERERS] FETCH_DELIVERER_SUCCESS',
  FETCH_DELIVERER_FAILURE = '[DELIVERERS] FETCH_DELIVERER_FAILURE',
  CREATE_DELIVERER = '[DELIVERERS] CREATE_DELIVERER',
  CREATE_DELIVERER_SUCCESS = '[DELIVERERS] CREATE_DELIVERER_SUCCESS',
  CREATE_DELIVERER_FAILURE = '[DELIVERERS] CREATE_DELIVERER_FAILURE',
  UPDATE_DELIVERER = '[DELIVERERS] UPDATE_DELIVERER',
  UPDATE_DELIVERER_SUCCESS = '[DELIVERERS] UPDATE_DELIVERER_SUCCESS',
  UPDATE_DELIVERER_FAILURE = '[DELIVERERS] UPDATE_DELIVERER_FAILURE',
  DELETE_DELIVERER = '[DELIVERERS] DELETE_DELIVERER',
  DELETE_DELIVERER_SUCCESS = '[DELIVERERS] DELETE_DELIVERER_SUCCESS',
  DELETE_DELIVERER_FAILURE = '[DELIVERERS] DELETE_DELIVERER_FAILURE',
}